Initial New Project with 2 axis scrolling
Character class in temporary and only for Demo purposes
LevelMap class requires changing to work with Atlas images
I added two levels to test the LevelMap Class
Cursor keys scroll the image & letters C and D change between levels

+++ crypt/src/com/crypt/Constant.java
@@ -3,7 +3,7 @@ package com.crypt;
 public class Constant 
 {
 	// Camera view screen height
-	public final static int SCREEN_HEIGHT = 768;
+	public final static float SCREEN_HEIGHT = 768;
 	
 	// Number of rows always required on screen
 	public final static int NUM_ROWS = 12;
